requirements.txt
opencv-python
opencv-python-headless
facedetector-py
filetype
numpy
pytest
pandas
requests

